Create Adding the transformer head 3
Adding the transformer head
Time to design a transformer head that could be used for classification tasks like sentiment analysis or categorization. You'll define a ClassifierHead class, create instances of the body and head, and pass a series of token IDs through them both to test that they work as expected.

Note: because this model has been trained yet, the outputs will be meaningless, but testing the code can process inputs and generate outputs in the form you expect is a good test.

Instructions 3/3

Complete the forward pass by passing input_sequence and src_mask to encoder, and the encoded output to classifier.